+++ research-site/src/App.css
@@ -1,40 +1,66 @@
+* {
+  margin: 0;
+  padding: 0;
+  box-sizing: border-box;
+}
+
+body {
+  font-family: Arial, sans-serif;
+}
+
 .container {
-  width: 80%;
-  margin: auto;
-  padding: 10px;
-  border: 1px solid black;
-  border: 10px;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  width: 100%;
+  padding: 20px;
 }
 
 .header {
+  width: 100%;
   text-align: center;
   margin-bottom: 20px;
+  font-size: 24px;
+  font-weight: bold;
 }
 
-.header h1 {
-  margin: 0;
-}
-
-.header nav ul li {
-  margin: 0 10px;
-}
-
-.main-content {
-  min-height: 400px;
+.main {
+  display: flex;
+  justify-content: space-between;
+  width: 100%;
   margin-bottom: 20px;
+}
+
+.box {
+  flex: 1;
+  margin: 0 10px;
+  padding: 20px;
+  border: 1px solid #000;
+  border-radius: 10px;
+  text-align: center;
+  background-color: #f0f0f0;
 }
 
 .footer {
-  border-top: 1px solid black;
-  padding-top: 10px;
-  text-align: center;
+  width: 100%;
+  text-align: left;
+  border: 1px solid #000;
+  padding: 10px;
+  border-radius: 10px;
+  background-color: #f0f0f0;
+}
+
+.footer p {
+  font-weight: bold;
+  margin: 0 0 10px 0;
 }
 
 .footer ul {
-  list-style-type: none;
   padding: 0;
+  margin: 0;
+  list-style-type: none;
 }
 
-.footer ul li {
-  margin: 5px 0;
+.footer li {
+  margin-bottom: 5px;
 }
