--- research-site/src/components/Footer.js
@@ -1,16 +1,0 @@
-import React from 'react';
-
-
-const Footer = () => {
-    return (
-        <footer className="footer">
-            <p>お知らせ</p>
-            <ul>
-                <li>殿堂入りするとここに掲載</li>
-                <li>いいね数の変動で形が変わります</li>
-            </ul>
-        </footer>
-    );
-};
-
-export default Footer;
