--- research-site/src/pages/Home.js
@@ -1,11 +1,0 @@
-import React from 'react';
-
-const Home = () => {
-    return (
-        <div>
-            <h2>ホームページコンテンツ</h2>
-        </div>
-    );
-};
-
-export default Home;
