--- research-site/src/pages/NewPost.js
@@ -1,11 +1,0 @@
-import React from 'react';
-
-const NewPost = () => {
-    return (
-        <div>
-            <h2>新規投稿ページ</h2>
-        </div>
-    );
-};
-
-export default NewPost;
