--- research-site/src/pages/Search.js
@@ -1,11 +1,0 @@
-import React from 'react';
-
-const Search = () => {
-    return (
-        <div>
-            <h2>検索ページ</h2>
-        </div>
-    );
-};
-
-export default Search;
